Add files via upload
BM3D (Block-matching and 3D filtering) can be said to be one of the best algorithms currently in effect. The idea of ​​this algorithm is similar to that of NL-Means. It is also a method of searching for similar blocks in the image, but it is much more complicated than NL-Means. Understanding NL-Means helps to understand the BM3D algorithm. The BM3D algorithm has a total of two major steps, which are divided into basic estimates (Step 1) and final estimates (Step 2): Basic Estimate and Final Estimate.
The main computation of this algorithm is still on the search and matching of similar blocks. In the case of similar blocks and search areas of the same size as NL-Means, the algorithm complexity of BM3D is higher than that of NL-Means. It should be approximately About 3 times of NL-Means. Students who dream of real-time processing can feel relieved.